+++ lib/src/repo.rs
@@ -588,7 +588,10 @@ impl MutableRepo {
         if let Some(current_checkout_id) = maybe_current_checkout_id {
             let current_checkout = self.store().get_commit(&current_checkout_id).unwrap();
             assert!(current_checkout.is_open(), "current checkout is closed");
-            if current_checkout.is_empty() && self.view().heads().contains(current_checkout.id()) {
+            if current_checkout.is_empty()
+                && current_checkout.description().is_empty()
+                && self.view().heads().contains(current_checkout.id())
+            {
                 // Abandon the checkout we're leaving if it's empty and a head commit
                 self.record_abandoned_commit(current_checkout_id);
             }